+++ lib/src/classes/EzRichText.dart
@@ -1,0 +1,243 @@
+/* empathetech_flutter_ui
+ * Copyright (c) 2023 Empathetech LLC. All rights reserved.
+ * See LICENSE for distribution and usage details.
+ */
+
+import 'dart:ui';
+import 'package:flutter/gestures.dart';
+import 'package:flutter/material.dart';
+
+class EzRichText extends StatelessWidget {
+  final String? text;
+  final TextSpan? textSpan;
+  final Key? key;
+  final FocusNode? focusNode;
+  final TextStyle? style;
+  final StrutStyle? strutStyle;
+  final TextAlign? textAlign;
+  final TextDirection? textDirection;
+  final double? textScaleFactor;
+  final bool showCursor;
+  final bool autofocus;
+  final int? minLines;
+  final int? maxLines;
+  final double cursorWidth;
+  final double? cursorHeight;
+  final Radius? cursorRadius;
+  final Color? cursorColor;
+  final BoxHeightStyle selectionHeightStyle;
+  final BoxWidthStyle selectionWidthStyle;
+  final DragStartBehavior dragStartBehavior;
+  final bool enableInteractiveSelection;
+  final TextSelectionControls? selectionControls;
+  final void Function()? onTap;
+  final ScrollPhysics scrollPhysics;
+  final String? semanticsLabel;
+  final TextHeightBehavior? textHeightBehavior;
+  final TextWidthBasis? textWidthBasis;
+  final void Function(TextSelection, SelectionChangedCause?)?
+      onSelectionChanged;
+  final EditableTextContextMenuBuilder? contextMenuBuilder;
+  final TextMagnifierConfiguration? magnifierConfiguration;
+
+  /// [SelectableText] wrapper with customized defaults
+  /// Takes in either [text] (originally [SelectableText.data])
+  /// or [textSpan] for a [SelectableText.rich] Widget
+  const EzRichText({
+    this.text,
+    this.textSpan,
+    this.key,
+    this.focusNode,
+    this.style,
+    this.strutStyle,
+    this.textAlign = TextAlign.center,
+    this.textDirection,
+    this.textScaleFactor,
+    this.showCursor = false,
+    this.autofocus = false,
+    this.minLines,
+    this.maxLines,
+    this.cursorWidth = 2.0,
+    this.cursorHeight,
+    this.cursorRadius,
+    this.cursorColor,
+    this.selectionHeightStyle = BoxHeightStyle.tight,
+    this.selectionWidthStyle = BoxWidthStyle.tight,
+    this.dragStartBehavior = DragStartBehavior.start,
+    this.enableInteractiveSelection = true,
+    this.selectionControls,
+    this.onTap,
+    this.scrollPhysics = const NeverScrollableScrollPhysics(),
+    this.semanticsLabel,
+    this.textHeightBehavior,
+    this.textWidthBasis,
+    this.onSelectionChanged,
+    this.contextMenuBuilder,
+    this.magnifierConfiguration,
+  }) : assert(
+          (text != null || textSpan != null) &&
+              !(text != null && textSpan != null),
+          'Either text or textSpan must be provided, but not both.',
+        );
+
+  @override
+  Widget build(BuildContext context) {
+    return (text == null)
+        ? SelectableText.rich(
+            textSpan!,
+            key: key,
+            focusNode: focusNode,
+            style: style,
+            strutStyle: strutStyle,
+            textAlign: textAlign,
+            textDirection: textDirection,
+            textScaleFactor: textScaleFactor,
+            showCursor: showCursor,
+            autofocus: autofocus,
+            minLines: minLines,
+            maxLines: maxLines,
+            cursorWidth: cursorWidth,
+            cursorHeight: cursorHeight,
+            cursorRadius: cursorRadius,
+            cursorColor: cursorColor,
+            selectionHeightStyle: selectionHeightStyle,
+            dragStartBehavior: dragStartBehavior,
+            enableInteractiveSelection: enableInteractiveSelection,
+            selectionControls: selectionControls,
+            onTap: onTap,
+            scrollPhysics: scrollPhysics,
+            semanticsLabel: semanticsLabel,
+            textHeightBehavior: textHeightBehavior,
+            textWidthBasis: textWidthBasis,
+            onSelectionChanged: onSelectionChanged,
+            magnifierConfiguration: magnifierConfiguration,
+          )
+        : SelectableText(
+            text!,
+            key: key,
+            focusNode: focusNode,
+            style: style,
+            strutStyle: strutStyle,
+            textAlign: textAlign,
+            textDirection: textDirection,
+            textScaleFactor: textScaleFactor,
+            showCursor: showCursor,
+            autofocus: autofocus,
+            minLines: minLines,
+            maxLines: maxLines,
+            cursorWidth: cursorWidth,
+            cursorHeight: cursorHeight,
+            cursorRadius: cursorRadius,
+            cursorColor: cursorColor,
+            selectionHeightStyle: selectionHeightStyle,
+            dragStartBehavior: dragStartBehavior,
+            enableInteractiveSelection: enableInteractiveSelection,
+            selectionControls: selectionControls,
+            onTap: onTap,
+            scrollPhysics: scrollPhysics,
+            semanticsLabel: semanticsLabel,
+            textHeightBehavior: textHeightBehavior,
+            textWidthBasis: textWidthBasis,
+            onSelectionChanged: onSelectionChanged,
+            magnifierConfiguration: magnifierConfiguration,
+          );
+  }
+}
+
+class EzRichTextBlock extends StatelessWidget {
+  final List<InlineSpan> children;
+  final Key? key;
+  final FocusNode? focusNode;
+  final TextStyle? style;
+  final StrutStyle? strutStyle;
+  final TextAlign? textAlign;
+  final TextDirection? textDirection;
+  final double? textScaleFactor;
+  final bool showCursor;
+  final bool autofocus;
+  final int? minLines;
+  final int? maxLines;
+  final double cursorWidth;
+  final double? cursorHeight;
+  final Radius? cursorRadius;
+  final Color? cursorColor;
+  final BoxHeightStyle selectionHeightStyle;
+  final BoxWidthStyle selectionWidthStyle;
+  final DragStartBehavior dragStartBehavior;
+  final bool enableInteractiveSelection;
+  final TextSelectionControls? selectionControls;
+  final void Function()? onTap;
+  final ScrollPhysics scrollPhysics;
+  final TextHeightBehavior? textHeightBehavior;
+  final TextWidthBasis? textWidthBasis;
+  final void Function(TextSelection, SelectionChangedCause?)?
+      onSelectionChanged;
+  final EditableTextContextMenuBuilder? contextMenuBuilder;
+  final TextMagnifierConfiguration? magnifierConfiguration;
+
+  /// [SelectableText.rich] wrapper with customized defaultsand  preconfigured semantics
+  /// Accepts [List] of [InlineSpan] rather than [TextSpan]
+  const EzRichTextBlock(
+    this.children, {
+    this.key,
+    this.focusNode,
+    this.style,
+    this.strutStyle,
+    this.textAlign = TextAlign.center,
+    this.textDirection,
+    this.textScaleFactor,
+    this.showCursor = false,
+    this.autofocus = false,
+    this.minLines,
+    this.maxLines,
+    this.cursorWidth = 2.0,
+    this.cursorHeight,
+    this.cursorRadius,
+    this.cursorColor,
+    this.selectionHeightStyle = BoxHeightStyle.tight,
+    this.selectionWidthStyle = BoxWidthStyle.tight,
+    this.dragStartBehavior = DragStartBehavior.start,
+    this.enableInteractiveSelection = true,
+    this.selectionControls,
+    this.onTap,
+    this.scrollPhysics = const NeverScrollableScrollPhysics(),
+    this.textHeightBehavior,
+    this.textWidthBasis,
+    this.onSelectionChanged,
+    this.contextMenuBuilder,
+    this.magnifierConfiguration,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    return SelectableText.rich(
+      TextSpan(children: children),
+      key: key,
+      focusNode: focusNode,
+      style: style,
+      strutStyle: strutStyle,
+      textAlign: textAlign,
+      textDirection: textDirection,
+      textScaleFactor: textScaleFactor,
+      showCursor: showCursor,
+      autofocus: autofocus,
+      minLines: minLines,
+      maxLines: maxLines,
+      cursorWidth: cursorWidth,
+      cursorHeight: cursorHeight,
+      cursorRadius: cursorRadius,
+      cursorColor: cursorColor,
+      selectionHeightStyle: selectionHeightStyle,
+      dragStartBehavior: dragStartBehavior,
+      enableInteractiveSelection: enableInteractiveSelection,
+      selectionControls: selectionControls,
+      onTap: onTap,
+      scrollPhysics: scrollPhysics,
+      semanticsLabel: 'FIX ME',
+      textHeightBehavior: textHeightBehavior,
+      textWidthBasis: textWidthBasis,
+      onSelectionChanged: onSelectionChanged,
+      magnifierConfiguration: magnifierConfiguration,
+    );
+  }
+}